refactor: move the isProjectPaper and asPaper functions from model/ directory to separate file
