+++ src/lib/components/composites/paper-components/PaperListEntry.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
     import PaperInfo from "$lib/components/composites/paper-components/PaperInfo.svelte";
-    import { asPaper } from "$lib/model/general";
     import UserAvatar from "$lib/components/composites/user-avatar/UserAvatar.svelte";
     import { goto } from "$app/navigation";
     import { PaperDecision } from "$lib/model/api/project";
@@ -9,6 +8,7 @@
     import { exhaustiveCheck } from "$lib/utils/common-helper";
     import { cn } from "$lib/utils/shadcn-helper";
     import type { PaperListEntryInterface } from "$lib/model/component-interfaces";
+    import { asPaper } from "$lib/utils/model-helper";
 
     type PaperListEntryProps = PaperListEntryInterface & {
         onClick?: () => void;
